refinements

+++ install-packages.sh
@@ -10,13 +10,13 @@ LOG_FILE=~/install-packages.log
 DATE=$(date '+%Y-%m-%d %H:%M:%S')
 DOTD=$(cd -- "$(dirname -- "${BASH_source[0]}" )" &> /dev/null && pwd)
 DIRS=()             # List of all directorys
-SLINK=()            # Source link path
+FILES=()            # Source link path
 DLINK=()            # Destination link path
 
-
+echo "Installing and setting up"
 # Check if it is a Debian distro
 if ! type -p apt > /dev/null; then
-    echo "APT is not installed, is it a Debian based distro?"
+    echo "   * ERROR: APT is not installed, is it a Debian based distro?"
     exit 1
 fi
 
@@ -29,19 +29,28 @@ done
 
 # Iterate each directory and source the install file
 for i in "${DIRS[@]}"; do 
+    echo -n "   "
 	source "$i/install.sh"
 
-    for (( j = 0; j < ${#SLINK[@]}; j++ )); do
-        rm -rf ${DLINK[$j]} > /dev/null 2>&1
-        ln -s ${DOTD}/${COMMAND}/${SLINK[$j]} ${DLINK[$j]}
-    done
-
-
+    # Check if command exists
 	if type -p $COMMAND > /dev/null; then
+	    echo "     * Installed "
 	    echo "$DATE: $COMMAND Installed " >> $LOG_FILE
 	else
+	    echo "     * Failed to install "
 	    echo "$DATE: $COMMAND failed to install" >> $LOG_FILE
 	fi
 
-    echo $DOTD
+    for (( j = 0; j < ${#FILES[@]}; j++ )); do
+        if test -f "${DOTD}/${COMMAND}/${FILES[$j]}"; then 
+            rm -rf ${DLINK[$j]} > /dev/null 2>&1
+            ln -s ${DOTD}/${COMMAND}/${FILES[$j]} ${DLINK[$j]}
+            echo "     * Symlink to ${DOTD}/${COMMAND}/${FILES[$j]} created"
+        else
+            echo "     * No config file found and no links was created"
+        fi
+    done
+
+    
+
 done
